# PHOENIX-6323 isRawFilterSupported() returns true for HBase 2.1 comapt…