--- a/phoenix-core/pom.xml
+++ b/phoenix-core/pom.xml
@@ -64,9 +64,9 @@
                     ("${hbase.compat.version}".equals("2.1.6")
                       &amp;&amp; hbaseMinor == 1
                       &amp;&amp; hbasePatch &gt;=6)
-                    || ("${hbase.compat.version}".equals("2.2.1")
+                    || ("${hbase.compat.version}".equals("2.2.5")
                       &amp;&amp; hbaseMinor == 2
-                      &amp;&amp; hbasePatch &gt;=1)
+                      &amp;&amp; hbasePatch &gt;=5)
                     || ("${hbase.compat.version}".equals("2.3.0")
                       &amp;&amp; hbaseMinor == 3
                       &amp;&amp; hbasePatch &gt;=0)