--- a/phoenix-hbase-compat-2.2.5/pom.xml
+++ b/phoenix-hbase-compat-2.2.5/pom.xml
@@ -26,12 +26,12 @@
     <version>5.1.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>phoenix-hbase-compat-2.2.1</artifactId>
-  <name>Phoenix Hbase 2.2.1 compatibility</name>
-  <description>Compatibility module for HBase 2.2.1+</description>
+  <artifactId>phoenix-hbase-compat-2.2.5</artifactId>
+  <name>Phoenix Hbase 2.2.5 compatibility</name>
+  <description>Compatibility module for HBase 2.2.5+</description>
 
   <properties>
-    <hbase22.compat.version>2.2.1</hbase22.compat.version>
+    <hbase22.compat.version>2.2.5</hbase22.compat.version>
   </properties>
 
   <dependencies>