--- a/phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/HbaseCompatCapabilities.java
+++ b/phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/HbaseCompatCapabilities.java
@@ -18,8 +18,6 @@
 
 package org.apache.phoenix.compat.hbase;
 
-import org.apache.hadoop.conf.Configuration;
-
 public class HbaseCompatCapabilities {
 
     public static boolean isMaxLookbackTimeSupported() {