--- a/pom.xml
+++ b/pom.xml
@@ -42,7 +42,7 @@
   <modules>
     <module>phoenix-hbase-compat-2.4.0</module>
     <module>phoenix-hbase-compat-2.3.0</module>
-    <module>phoenix-hbase-compat-2.2.1</module>
+    <module>phoenix-hbase-compat-2.2.5</module>
     <module>phoenix-hbase-compat-2.1.6</module>
     <module>phoenix-core</module>
     <module>phoenix-pherf</module>
@@ -671,7 +671,7 @@
       </dependency>
       <dependency>
         <groupId>org.apache.phoenix</groupId>
-        <artifactId>phoenix-hbase-compat-2.2.1</artifactId>
+        <artifactId>phoenix-hbase-compat-2.2.5</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -1549,7 +1549,7 @@
     </profile>
     <profile>
       <!-- PHOENIX-5993 This won't work with the public HBase artifacts -->
-      <id>phoenix-hbase-compat-2.2.1</id>
+      <id>phoenix-hbase-compat-2.2.5</id>
       <activation>
         <property>
           <name>hbase.profile</name>
@@ -1558,7 +1558,7 @@
       </activation>
       <properties>
         <hbase.profile>2.2</hbase.profile>
-        <hbase.compat.version>2.2.1</hbase.compat.version>
+        <hbase.compat.version>2.2.5</hbase.compat.version>
         <hbase.version>${hbase-2.2.runtime.version}</hbase.version>
       </properties>
     </profile>