… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatDelegateHTable.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatOmidTransactionTable.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatPermissionUtil.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatPhoenixRpcScheduler.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatSteppingSplitPolicy.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatStoreFileReader.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/CompatUtil.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/OffsetCell.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/coprocessor/CompatBaseScannerRegionObserver.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/coprocessor/CompatIndexRegionObserver.java
… phoenix-hbase-compat-2.2.5/src/main/java/org/apache/phoenix/compat/hbase/test/DelegateCell.java