@stoty 2021-01-18
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| +0 :ok: |  reexec  |   1m 19s |  Docker mode activated.  |
||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| +1 :green_heart: |  hbaseanti  |   0m  0s |  Patch does not have any anti-patterns.  |
| +1 :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| -1 :x: |  test4tests  |   0m  0s |  The patch doesn't appear to include any new or modified tests. Please justify why no new tests are needed for this patch. Also please list what manual steps were performed to verify this patch.  |
||| _ master Compile Tests _ |
| +0 :ok: |  mvndep  |   4m 22s |  Maven dependency ordering for branch  |
| +1 :green_heart: |  mvninstall  |  12m 42s |  master passed  |
| +0 |  hbaserecompile  |  24m  5s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 37s |  master passed  |
| +1 :green_heart: |  checkstyle  |   0m 38s |  master passed  |
| +1 :green_heart: |  javadoc  |   0m 35s |  master passed  |
| +0 :ok: |  spotbugs  |   0m 35s |  phoenix-hbase-compat-2.2.1 in master has 1 extant spotbugs warnings.  |
| +0 :ok: |  spotbugs  |   0m 31s |  phoenix-hbase-compat-2.1.6 in master has 1 extant spotbugs warnings.  |
||| _ Patch Compile Tests _ |
| +0 :ok: |  mvndep  |   0m 11s |  Maven dependency ordering for patch  |
| +1 :green_heart: |  mvninstall  |   9m 43s |  the patch passed  |
| +0 |  hbaserecompile  |  16m 47s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 36s |  the patch passed  |
| +1 :green_heart: |  javac  |   1m 36s |  the patch passed  |
| -1 :x: |  checkstyle  |   0m 39s |  root: The patch generated 2 new + 15 unchanged - 3 fixed = 17 total (was 18)  |
| +1 :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| +1 :green_heart: |  javadoc  |   0m 33s |  the patch passed  |
| +1 :green_heart: |  spotbugs  |   1m 27s |  the patch passed  |
||| _ Other Tests _ |
| -1 :x: |  unit  | 124m 28s |  root in the patch failed.  |
| +1 :green_heart: |  asflicense  |   0m 55s |  The patch does not generate ASF License warnings.  |
|  |   | 179m  7s |   |


| Reason | Tests |
|-------:|:------|
| Failed junit tests | phoenix.end2end.ProductMetricsIT |
|   | phoenix.end2end.UpsertSelectIT |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.41 ServerAPI=1.41 base: https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/1/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/phoenix/pull/1095 |
| Optional Tests | dupname asflicense javac javadoc unit spotbugs hbaserebuild hbaseanti checkstyle compile |
| uname | Linux 0ee514011613 4.15.0-128-generic #131-Ubuntu SMP Wed Dec 9 06:57:35 UTC 2020 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev/phoenix-personality.sh |
| git revision | master / 50df995 |
| Default Java | Private Build-1.8.0_242-8u242-b08-0ubuntu3~16.04-b08 |
| checkstyle | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/1/artifact/yetus-general-check/output/diff-checkstyle-root.txt |
| unit | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/1/artifact/yetus-general-check/output/patch-unit-root.txt |
|  Test Results | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/1/testReport/ |
| Max. process+thread count | 9525 (vs. ulimit of 30000) |
| modules | C: phoenix-hbase-compat-2.2.1 phoenix-hbase-compat-2.1.6 U: . |
| Console output | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/1/console |
| versions | git=2.7.4 maven=3.3.9 spotbugs=4.1.3 |
| Powered by | Apache Yetus 0.12.0 https://yetus.apache.org |


This message was automatically generated.