@gjacoby126 2021-01-20
@stoty - it would be very good if we could support the raw filter scans in the HBase 2.2 branch. They're required to fix a serious bug where building an index on a view can't apply the view's WHERE clause filter condition properly to the index, and so you wind up indexing the entire table in each view index, wasting lots of storage by creating tons of corrupt index rows (that are never read). @abhishek-chouhan discovered and fixed it in PHOENIX-5932. 

Does requiring 2.2.5 on the 2.2 branch create difficulties?