@stoty 2021-01-20
Thanks for the explanation, @gjacoby126.

For reference, these are our options:

* bump the minimum version to 2.2.5 
* split the 2.2 pofile into two
* or make isRawFilterSupported() dynamic depending on the detected HBase version

* bumping the version is the simplest, but we lock out some users from Phoenix.
* splitting the profile adds a bit of bloat, and possible confusion
* making isRawFilterSupported() dynamic is easy, but will cause divergent behaviour for different HBase versions.

I'm going to update the PR to bump the version requirement, but LMK if you prefer one of the other versions.