@stoty 2021-01-20
New version requires HBase 2.2.5