@stoty 2021-01-20
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| +0 :ok: |  reexec  |   1m 17s |  Docker mode activated.  |
||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| +1 :green_heart: |  hbaseanti  |   0m  0s |  Patch does not have any anti-patterns.  |
| +1 :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| +1 :green_heart: |  test4tests  |   0m  0s |  The patch appears to include 2 new or modified test files.  |
||| _ master Compile Tests _ |
| +0 :ok: |  mvndep  |   4m 37s |  Maven dependency ordering for branch  |
| +1 :green_heart: |  mvninstall  |  13m 12s |  master passed  |
| +0 |  hbaserecompile  |  26m  3s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 43s |  master passed  |
| +1 :green_heart: |  checkstyle  |   0m 53s |  master passed  |
| +1 :green_heart: |  javadoc  |   3m  9s |  master passed  |
| +0 :ok: |  spotbugs  |   0m 35s |  phoenix-hbase-compat-2.2.1 in master has 1 extant spotbugs warnings.  |
| +0 :ok: |  spotbugs  |   0m 35s |  phoenix-hbase-compat-2.1.6 in master has 1 extant spotbugs warnings.  |
| +0 :ok: |  spotbugs  |   5m  1s |  root in master has 1004 extant spotbugs warnings.  |
| +0 :ok: |  spotbugs  |   3m 27s |  phoenix-core in master has 955 extant spotbugs warnings.  |
||| _ Patch Compile Tests _ |
| +0 :ok: |  mvndep  |   0m 33s |  Maven dependency ordering for patch  |
| +1 :green_heart: |  mvninstall  |  11m 39s |  the patch passed  |
| +0 |  hbaserecompile  |  20m 16s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 39s |  the patch passed  |
| +1 :green_heart: |  javac  |   1m 39s |  the patch passed  |
| +1 :green_heart: |  checkstyle  |   0m 42s |  root: The patch generated 0 new + 215 unchanged - 2 fixed = 215 total (was 217)  |
| +1 :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| +1 :green_heart: |  xml  |   0m  3s |  The patch has no ill-formed XML file.  |
| +1 :green_heart: |  javadoc  |   2m 52s |  the patch passed  |
| +1 :green_heart: |  spotbugs  |  10m 42s |  the patch passed  |
||| _ Other Tests _ |
| -1 :x: |  unit  | 192m 54s |  root in the patch failed.  |
| -1 :x: |  asflicense  |   2m 35s |  The patch generated 610 ASF License warnings.  |
|  |   | 280m 28s |   |


| Reason | Tests |
|-------:|:------|
| Failed junit tests | phoenix.end2end.index.GlobalMutableNonTxIndexWithLazyPostBatchWriteIT |
|   | phoenix.end2end.index.txn.RollbackIT |
|   | phoenix.tx.ParameterizedTransactionIT |
|   | phoenix.tx.FlappingTransactionIT |
|   | phoenix.end2end.QueryIT |
|   | phoenix.end2end.salted.SaltedTableUpsertSelectIT |
|   | phoenix.replication.SystemCatalogWALEntryFilterIT |
|   | phoenix.end2end.index.ImmutableIndexExtendedIT |
|   | phoenix.end2end.ViewTTLIT |
|   | phoenix.rpc.UpdateCacheConnectionLevelPropIT |
|   | phoenix.util.IndexScrutinyIT |
|   | phoenix.end2end.UpsertSelectIT |
|   | phoenix.iterate.PhoenixQueryTimeoutIT |
|   | phoenix.end2end.join.HashJoinLocalIndexIT |
|   | phoenix.rpc.UpdateCacheIT |
|   | phoenix.end2end.salted.SaltedTableVarLengthRowKeyIT |
|   | phoenix.end2end.index.IndexCoprocIT |
|   | phoenix.end2end.join.WhereOptimizerForJoinFiltersIT |
|   | phoenix.end2end.index.MutableIndexSplitForwardScanIT |
|   | phoenix.iterate.RoundRobinResultIteratorIT |
|   | phoenix.end2end.salted.SaltedTableIT |
|   | phoenix.end2end.join.SortMergeJoinNoIndexIT |
|   | phoenix.end2end.index.SingleCellIndexIT |
|   | phoenix.end2end.join.HashJoinPersistentCacheIT |
|   | phoenix.end2end.index.LocalMutableNonTxIndexIT |
|   | phoenix.end2end.join.SubqueryIT |
|   | phoenix.end2end.index.SaltedIndexIT |
|   | phoenix.end2end.index.MutableIndexIT |
|   | phoenix.end2end.RowValueConstructorOffsetIT |
|   | phoenix.end2end.DisableLocalIndexIT |
|   | phoenix.end2end.ToCharFunctionIT |
|   | phoenix.end2end.join.SortMergeJoinLocalIndexIT |
|   | phoenix.end2end.join.HashJoinGlobalIndexIT |
|   | phoenix.tx.TxCheckpointIT |
|   | phoenix.end2end.join.HashJoinNoIndexIT |
|   | phoenix.iterate.ChunkedResultIteratorIT |
|   | phoenix.end2end.index.InvalidIndexStateClientSideIT |
|   | phoenix.end2end.join.SortMergeJoinGlobalIndexIT |
|   | phoenix.tx.TransactionIT |
|   | phoenix.end2end.QueryExecWithoutSCNIT |
|   | phoenix.end2end.join.SubqueryUsingSortMergeJoinIT |
|   | phoenix.end2end.index.GlobalMutableTxIndexIT |
|   | phoenix.end2end.PrimitiveTypeIT |
|   | phoenix.end2end.index.MutableIndexSplitReverseScanIT |
|   | phoenix.end2end.join.HashJoinMoreIT |
|   | phoenix.monitoring.IndexMetricsIT |
|   | phoenix.end2end.index.txn.MutableRollbackIT |
|   | phoenix.end2end.UpsertWithSCNIT |
|   | phoenix.trace.PhoenixTableMetricsWriterIT |
|   | phoenix.end2end.OnDuplicateKeyIT |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.41 ServerAPI=1.41 base: https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/2/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/phoenix/pull/1095 |
| Optional Tests | dupname asflicense javac javadoc unit xml compile spotbugs hbaserebuild hbaseanti checkstyle |
| uname | Linux 8cd070844151 4.15.0-128-generic #131-Ubuntu SMP Wed Dec 9 06:57:35 UTC 2020 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev/phoenix-personality.sh |
| git revision | master / 772ff2c |
| Default Java | Private Build-1.8.0_242-8u242-b08-0ubuntu3~16.04-b08 |
| unit | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/2/artifact/yetus-general-check/output/patch-unit-root.txt |
|  Test Results | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/2/testReport/ |
| asflicense | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/2/artifact/yetus-general-check/output/patch-asflicense-problems.txt |
| Max. process+thread count | 8336 (vs. ulimit of 30000) |
| modules | C: phoenix-hbase-compat-2.2.5 phoenix-hbase-compat-2.1.6 phoenix-core . U: . |
| Console output | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1095/2/console |
| versions | git=2.7.4 maven=3.3.9 spotbugs=4.1.3 |
| Powered by | Apache Yetus 0.12.0 https://yetus.apache.org |


This message was automatically generated.